fix: 🐛 use of the package in browsers
